pyproject.toml
[project]
name = "heatingpi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "gpiozero>=2.0.1",
    "pyserial>=3.5",
    "requests>=2.32.3",
]
